added windows

+++ settingsWindow.xaml.cs
@@ -43,54 +43,26 @@ namespace COE131L
             loggedUser = userId;
             //ADD COMBO BOX ELEMENTS 
            
-            dayList.Add("01");
-            dayList.Add("02");
-            dayList.Add("03");
-            dayList.Add("04");
-            dayList.Add("05");
-            dayList.Add("06");
-            dayList.Add("07");
-            dayList.Add("08");
-            dayList.Add("09");
-            dayList.Add("10");
-            dayList.Add("11");
-            dayList.Add("12");
-            dayList.Add("13");
-            dayList.Add("14");
-            dayList.Add("15");
-            dayList.Add("16");
-            dayList.Add("17");
-            dayList.Add("18");
-            dayList.Add("19");
-            dayList.Add("20");
-            dayList.Add("21");
-            dayList.Add("22");
-            dayList.Add("23");
-            dayList.Add("24");
-            dayList.Add("25");
-            dayList.Add("26");
-            dayList.Add("27");
-            dayList.Add("28");
-            dayList.Add("29");
-            dayList.Add("30");
+            dayList.Add("01");dayList.Add("02");dayList.Add("03");
+            dayList.Add("04");dayList.Add("05");dayList.Add("06");
+            dayList.Add("07");dayList.Add("08");dayList.Add("09");
+            dayList.Add("10");dayList.Add("11");dayList.Add("12");
+            dayList.Add("13");dayList.Add("14");dayList.Add("15");
+            dayList.Add("16");dayList.Add("17");dayList.Add("18");
+            dayList.Add("19");dayList.Add("20");dayList.Add("21");
+            dayList.Add("22");dayList.Add("23");dayList.Add("24");
+            dayList.Add("25");dayList.Add("26");dayList.Add("27");
+            dayList.Add("28");dayList.Add("29");dayList.Add("30");
             dayList.Add("31");
 
             comboDay.ItemsSource = dayList;
 
-            
-                monthList.Add("01");
-                monthList.Add("02");
-                monthList.Add("03");
-                monthList.Add("04");
-                monthList.Add("05");
-                monthList.Add("06");
-                monthList.Add("07");
-                monthList.Add("08");
-                monthList.Add("09");
-                monthList.Add("10");
-                monthList.Add("11");
-                monthList.Add("12");
-            
+            monthList.Add("01");monthList.Add("02");
+            monthList.Add("03"); monthList.Add("04");
+            monthList.Add("05");monthList.Add("06");
+            monthList.Add("07");monthList.Add("08");
+            monthList.Add("09");monthList.Add("10");
+            monthList.Add("11"); monthList.Add("12");
             this.comboMonth.ItemsSource = monthList;
 
         }
@@ -100,33 +72,54 @@ namespace COE131L
             string typeName = this.textboxName.Text;
             string typeModel = this.textboxModel.Text;
 
+            //ITEM TYPE ADD MODE
             if (this.addRadio.IsChecked == true)
             {
-                if (Database.addnewtype(typeName, typeModel) == true)
+                if (string.IsNullOrWhiteSpace(typeName) == false || string.IsNullOrWhiteSpace(typeModel) == false)
                 {
-                    //MEANS THAT THE INSERT FAILED MAKEA PROMPT OR MESSAGE
-                    MessageBox.Show("Insert of new type failed! Please try again.", "Incorrect input", MessageBoxButton.OK, MessageBoxImage.Exclamation);
+                    if (Database.addnewtype(typeName, typeModel) == true)
+                    {
+                        //MEANS THAT THE INSERT FAILED MAKEA PROMPT OR MESSAGE
+                        MessageBox.Show("Insert of new type failed! Please try again.", "Incorrect input", MessageBoxButton.OK, MessageBoxImage.Exclamation);
+                    }
+                    else
+                    {
+                       //MEANS THAT THE INSERT SUCCESSFUL MAKE A PROMPT OR MESSAGE SUCCESSFUL INSERT
+                       MessageBox.Show("Insert of new type success!", "Successful Insert", MessageBoxButton.OK, MessageBoxImage.Information);
+                       this.comboType.ItemsSource = Database.getItemtypes();
+                    }
                 }
                 else
                 {
-                    //MEANS THAT THE INSERT SUCCESSFUL MAKE A PROMPT OR MESSAGE SUCCESSFUL INSERT
-                    MessageBox.Show("Insert of new type success!", "Successful Insert", MessageBoxButton.OK, MessageBoxImage.Information);
-                    
+                    MessageBox.Show("Please enter all required information!","Insufficient input",MessageBoxButton.OK,MessageBoxImage.Error);
                 }
-            }
-            else if(this.removeRadio.IsChecked == true)
+            }//ITEM TYPE REMOVE MODE
+            else if (this.removeRadio.IsChecked == true)
             {
-                if(Database.removetype(typeName,typeModel) == false)
+                if (string.IsNullOrWhiteSpace(typeName) == false && string.IsNullOrWhiteSpace(typeModel) == false)
                 {
-                    //REMOVING OF THE TYPE IS NOT SUCCESSFUL MAKEA PROMOT TO THE USER
-                    MessageBox.Show("Removal of a new type failed! Please try again.", "Incorrect input", MessageBoxButton.OK, MessageBoxImage.Exclamation);
+                    if (Database.removetype(typeName, typeModel) == false)
+                    {
+                        //REMOVING OF THE TYPE IS NOT SUCCESSFUL MAKEA PROMOT TO THE USER
+                        MessageBox.Show("Removal of a new type failed! Please try again.", "Incorrect input", MessageBoxButton.OK, MessageBoxImage.Exclamation);
+                    }
+                    else
+                    {
+                        MessageBox.Show("Removal of a new type success!", "Successful removal", MessageBoxButton.OK, MessageBoxImage.Information);
+                        this.comboType.ItemsSource = Database.getItemtypes();
+                    }
                 }
                 else
                 {
-                    MessageBox.Show("Removal of a new type success!", "Successful removal", MessageBoxButton.OK, MessageBoxImage.Information);
+                    MessageBox.Show("Input all required information!","Insufficient Information",MessageBoxButton.OK,MessageBoxImage.Information);
                 }
             }
-
+           else
+           {
+                MessageBox.Show("Please select a mode!", "Select Mode", MessageBoxButton.OK, MessageBoxImage.Exclamation);    
+           }
+            
+        
         }
 
         private void typeRadio_Checked(object sender, RoutedEventArgs e)
@@ -166,46 +159,60 @@ namespace COE131L
             //ADD SELECTED
             if (this.radioEqauipadd.IsChecked == true)
             {
-              
+                
+                    item newItem = new item();
+                    string date = this.comboDay.SelectedItem.ToString() + "/" + this.comboMonth.SelectedItem.ToString() + "/" + this.textboxYear.Text;
+                    DateTime deldate = DateTime.ParseExact(date, "dd/MM/yyyy", null);
+                    DateTime datedecom = deldate.AddMonths(Int32.Parse(this.textblockMonth.Text));
+                    datedecom = datedecom.AddYears(Int32.Parse(this.textblockYear.Text));
 
-                item newItem = new item();
-                string date = this.comboDay.SelectedItem.ToString() + "/" + this.comboMonth.SelectedItem.ToString() + "/" + this.textboxYear.Text;
-                DateTime deldate = DateTime.ParseExact(date, "dd/MM/yyyy", null);
-                DateTime datedecom = deldate.AddMonths(Int32.Parse(this.textblockMonth.Text));
-                datedecom = datedecom.AddYears(Int32.Parse(this.textblockYear.Text));
+                    newItem.serialNumber = Int32.Parse(this.textboxSerial.Text);
+                    newItem.itemType = this.comboType.SelectedIndex + 1;
+                    newItem.addedby = loggedUser;
+                    newItem.statusId = this.itemStat;
+                    newItem.conditionId = this.condId;
+                    newItem.supplier = this.textboxSupplier.Text;
+                    newItem.datedelivered = deldate.ToString("dd/MM/yyyy");
+                    newItem.datedecomm = datedecom.ToString("dd/MM/yyyy");
+                    newItem.model = this.comboModel.SelectedItem.ToString();
 
-                newItem.serialNumber = Int32.Parse(this.textboxSerial.Text);
-                newItem.itemType = this.comboType.SelectedIndex + 1;
-                newItem.addedby = loggedUser;
-                newItem.statusId = this.itemStat;
-                newItem.conditionId = this.condId;
-                newItem.supplier = this.textboxSupplier.Text;
-                newItem.datedelivered = deldate.ToString("dd/MM/yyyy");
-                newItem.datedecomm = datedecom.ToString("dd/MM/yyyy");
-                newItem.model = this.comboModel.SelectedItem.ToString();
+                    if (Database.addItem(newItem) == true) //FAILED TO ADD ITEM TO THE RECORD 
+                    {
+                        MessageBox.Show("Item already exists in the record! Please enter other information.", "Existing item", MessageBoxButton.OK, MessageBoxImage.Error);
+                        this.textboxModel.Clear();
+                        this.textboxName.Clear();
+                        this.addRadio.IsChecked = false;
+                        this.removeRadio.IsChecked = false;
+                    }
+                    else //SUCCESSFULLY ADDED
+                    {
+                        MessageBox.Show("Item is added to the record!", "Item added", MessageBoxButton.OK, MessageBoxImage.Information);
+                        wind.loadDatagrid();
+                        //CLEAR ALL THE CONTENT OF THE EQUIPMENT PART
+                        this.textboxModel.Clear();
+                        this.textboxName.Clear();
+                        this.addRadio.IsChecked = false;
+                        this.removeRadio.IsChecked = false;
 
-                if (Database.addItem(newItem) == true) //FAILED TO ADD ITEM TO THE RECORD 
-                {
-                    MessageBox.Show("Item already exists in the record! Please enter other information.", "Existing item", MessageBoxButton.OK, MessageBoxImage.Error);
-                }
-                else //SUCCESSFULLY ADDED
-                {
-                    MessageBox.Show("Item is added to the record!", "Item added", MessageBoxButton.OK, MessageBoxImage.Information);
-                    wind.loadDatagrid();
-                    //CLEAR ALL THE CONTENT OF THE RECORD  
-                }
+                    }
+                    
+               
             }
             //remove selected
             else if(this.raduiEquiprem.IsChecked == true)
             {
-                if(Database.removeItem(Int32.Parse(this.textboxSerial.Text)) == true)//ITEM IS REMOVED 
-                {
-                    MessageBox.Show("Item is successfully removed from the record.", "Item Removed", MessageBoxButton.OK, MessageBoxImage.Information);
-                }
-                else
-                {
-                    MessageBox.Show("Item failed to be removed from the record.", "Removal Failed", MessageBoxButton.OK, MessageBoxImage.Error);
-                }
+
+                    if (Database.removeItem(Int32.Parse(this.textboxSerial.Text)) == true)//ITEM IS REMOVED 
+                    {
+                        MessageBox.Show("Item is successfully removed from the record.", "Item Removed", MessageBoxButton.OK, MessageBoxImage.Information);
+                        wind.loadDatagrid();
+                    }
+                    else
+                    {
+                        MessageBox.Show("Item failed to be removed from the record.", "Removal Failed", MessageBoxButton.OK, MessageBoxImage.Error);
+                    }
+                
+                
             }
             // none selected
             else
